+++ yang/yang-parser-rfc7950/src/main/java/org/opendaylight/yangtools/yang/parser/rfc7950/stmt/position/AbstractPositionEffectiveStatement.java
@@ -7,12 +7,13 @@
  */
 package org.opendaylight.yangtools.yang.parser.rfc7950.stmt.position;
 
+import org.opendaylight.yangtools.yang.common.Uint32;
 import org.opendaylight.yangtools.yang.model.api.stmt.PositionEffectiveStatement;
 import org.opendaylight.yangtools.yang.model.api.stmt.PositionStatement;
 import org.opendaylight.yangtools.yang.parser.rfc7950.stmt.AbstractDeclaredEffectiveStatement;
 
 abstract class AbstractPositionEffectiveStatement
-        extends AbstractDeclaredEffectiveStatement.DefaultArgument<Long, PositionStatement>
+        extends AbstractDeclaredEffectiveStatement.DefaultArgument<Uint32, PositionStatement>
         implements PositionEffectiveStatement {
     AbstractPositionEffectiveStatement(final PositionStatement declared) {
         super(declared);
